src/requirements.txt
blinker==1.6.2
click==8.1.6
colorama==0.4.6
dnspython==2.4.2
Flask==2.3.2
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
git+https://github.com/CIAT-DAPA/lswms_models
mongoengine==0.26.0
pymongo==4.4.1
python-dotenv==1.0.0
werkzeug==2.3.7
zipp==3.16.2
mongomock==4.1.2